Revert "Merge pull request #3 from massif-01/feature/ui-optimization"
This reverts commit 7f29d5bbd96ce0a0d20cea88a95fa0eded00e91b, reversing
changes made to 73eea56fed45b1aadabb677f14d1828055ea224c.

+++ components/ts_ota/src/ts_ota_rollback.c
@@ -167,7 +167,7 @@ esp_err_t ts_ota_save_update_time(void)
     ret = nvs_commit(handle);
     nvs_close(handle);
 
-    ESP_LOGI(TAG, "Update #%u recorded at %lld", (unsigned int)count, (long long)now);
+    ESP_LOGI(TAG, "Update #%lu recorded at %ld", count, now);
     return ret;
 }
